chore: Modal 폴더를 portal 폴더로 변경
