+++ src/feature/portal/components/style/modal.scss
@@ -1,4 +1,4 @@
-.modal-outer-container {
+.modal-container {
     display: flex;
     justify-content: center;
     align-items: center;
