+++ src/routes/profile/$mutableId/mutualFollowers.tsx
@@ -14,7 +14,7 @@ import { createFileRoute } from '@tanstack/react-router';
 /* utils */
 import { getFirstNodeOfCollection } from '$lib/utils/graphql';
 /* components */
-import ModalContainer from '$feature/Modal/components/ModalContainer';
+import Modal from '$feature/portal/components/Modal';
 import MiniProfileIterator from '$feature/Profile/MiniProfileIterator';
 
 export const Route = createFileRoute('/profile/$mutableId/mutualFollowers')({
@@ -49,8 +49,8 @@ function MutualFollowingModalComponent(props: PropsWithStatus<ModalMiniProfilePr
     }
 
     return (
-        <ModalContainer title={title} handleClickClose={handleClickClose}>
+        <Modal title={title} handleClickClose={handleClickClose}>
             <MiniProfileIterator {...props}/>
-        </ModalContainer>
+        </Modal>
     )
 }
